Separate loss and EMA
