+++ mixmatch/utils/loss.py
@@ -1,6 +1,5 @@
 import numpy as np
 import torch
-import torch.nn as nn
 import torch.nn.parallel
 from torch.nn.functional import cross_entropy
 
@@ -36,31 +35,3 @@ class SemiLoss(object):
             l_u,
             lambda_u * self.linear_rampup(epoch, epochs),
         )
-
-
-class WeightEMA(object):
-    def __init__(
-        self,
-        model: nn.Module,
-        ema_model: nn.Module,
-        alpha: float = 0.999,
-        lr: float = 0.002,
-    ):
-        self.model = model
-        self.ema_model = ema_model
-        self.alpha = alpha
-        self.params = list(model.state_dict().values())
-        self.ema_params = list(ema_model.state_dict().values())
-        self.wd = 0.02 * lr
-
-        for param, ema_param in zip(self.params, self.ema_params):
-            param.data.copy_(ema_param.data)
-
-    def step(self):
-        one_minus_alpha = 1.0 - self.alpha
-        for param, ema_param in zip(self.params, self.ema_params):
-            if ema_param.dtype == torch.float32:
-                ema_param.mul_(self.alpha)
-                ema_param.add_(param * one_minus_alpha)
-                # customized weight decay
-                param.mul_(1 - self.wd)
